+++ camo/backend/__init__.py
@@ -1,3 +1,4 @@
+from .conditional_independence_test import *
 from .directed_graph import *
 from .directed_markov_graph import *
 from .graph import *