+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Added Back-door and Front-door paths discover function
 * Added ACE confidence interval with bootstrap
+* Added conditional independence tests
 
 ### Changed
 ### Removed